pakage name 수정 (model -> entity)
