+++ src/main/java/com/sparta/project/controller/AuthController.java
@@ -3,10 +3,8 @@ package com.sparta.project.controller;
 import com.sparta.project.dto.user.LoginRequestDto;
 import com.sparta.project.dto.user.LoginResponseDto;
 import com.sparta.project.dto.user.UserRequestDto;
-import com.sparta.project.model.User;
 import com.sparta.project.repository.UserRepository;
 import com.sparta.project.service.AuthService;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
