+++ src/main/java/com/sparta/project/controller/UserController.java
@@ -1,11 +1,8 @@
 package com.sparta.project.controller;
 
-import com.sparta.project.dto.match.MatchResponseDto;
 import com.sparta.project.dto.message.MessageResponseDto;
 import com.sparta.project.dto.user.EvaluationDto;
 import com.sparta.project.dto.user.MyPageResponseDto;
-import com.sparta.project.model.User;
-import com.sparta.project.service.AuthService;
 import com.sparta.project.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
