+++ src/main/java/com/sparta/project/dto/message/MessageRequestDto.java
@@ -1,7 +1,7 @@
 package com.sparta.project.dto.message;
 
-import com.sparta.project.model.Match;
-import com.sparta.project.model.User;
+import com.sparta.project.entity.Match;
+import com.sparta.project.entity.User;
 import lombok.Getter;
 import lombok.Setter;
