+++ src/main/java/com/sparta/project/dto/token/TokenRequestDto.java
@@ -1,9 +1,7 @@
 package com.sparta.project.dto.token;
 
-import com.sparta.project.model.User;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
