+++ src/main/java/com/sparta/project/dto/user/MyPageResponseDto.java
@@ -1,7 +1,6 @@
 package com.sparta.project.dto.user;
 
 import com.sparta.project.dto.match.MatchResponseDto;
-import com.sparta.project.model.Match;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -11,7 +10,7 @@ import java.util.List;
 @Getter
 @Setter
 @Builder
-public class MyPageResponseDto {
+public class  MyPageResponseDto {
 
     private String nickname;
     private Long score;
