+++ src/main/java/com/sparta/project/entity/RequestUserList.java
@@ -1,4 +1,4 @@
-package com.sparta.project.model;
+package com.sparta.project.entity;
 
 import com.sparta.project.dto.user.InviteResponseDto;
 import lombok.Getter;
