+++ src/main/java/com/sparta/project/repository/BowlingRepository.java
@@ -1,8 +1,8 @@
 package com.sparta.project.repository;
 
-import com.sparta.project.model.Bowling;
-import com.sparta.project.model.Match;
-import com.sparta.project.model.User;
+import com.sparta.project.entity.Bowling;
+import com.sparta.project.entity.Match;
+import com.sparta.project.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
