+++ src/main/java/com/sparta/project/repository/EvaluationRepository.java
@@ -1,6 +1,6 @@
 package com.sparta.project.repository;
 
-import com.sparta.project.model.Evaluation;
+import com.sparta.project.entity.Evaluation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
