+++ src/main/java/com/sparta/project/repository/MatchRepository.java
@@ -1,6 +1,6 @@
 package com.sparta.project.repository;
 
-import com.sparta.project.model.Match;
+import com.sparta.project.entity.Match;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
