+++ src/main/java/com/sparta/project/repository/MessageRepository.java
@@ -1,8 +1,7 @@
 package com.sparta.project.repository;
 
-import com.sparta.project.dto.message.MessageResponseDto;
-import com.sparta.project.model.Match;
-import com.sparta.project.model.Message;
+import com.sparta.project.entity.Match;
+import com.sparta.project.entity.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
