+++ src/main/java/com/sparta/project/repository/RequestUserListRepository.java
@@ -1,7 +1,7 @@
 package com.sparta.project.repository;
 
-import com.sparta.project.model.Match;
-import com.sparta.project.model.RequestUserList;
+import com.sparta.project.entity.Match;
+import com.sparta.project.entity.RequestUserList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
