+++ src/main/java/com/sparta/project/repository/UserListInMatchRepository.java
@@ -1,8 +1,8 @@
 package com.sparta.project.repository;
 
-import com.sparta.project.model.Match;
-import com.sparta.project.model.User;
-import com.sparta.project.model.UserListInMatch;
+import com.sparta.project.entity.Match;
+import com.sparta.project.entity.User;
+import com.sparta.project.entity.UserListInMatch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
