+++ src/main/java/com/sparta/project/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.sparta.project.repository;
 
 
-import com.sparta.project.model.User;
+import com.sparta.project.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
