+++ src/main/java/com/sparta/project/service/AuthService.java
@@ -5,8 +5,8 @@ import com.sparta.project.dto.user.LoginRequestDto;
 import com.sparta.project.dto.user.LoginResponseDto;
 import com.sparta.project.dto.user.UserRequestDto;
 import com.sparta.project.dto.user.UserResponseDto;
-import com.sparta.project.model.RefreshToken;
-import com.sparta.project.model.User;
+import com.sparta.project.entity.RefreshToken;
+import com.sparta.project.entity.User;
 import com.sparta.project.repository.RefreshTokenRepository;
 import com.sparta.project.repository.UserRepository;
 import com.sparta.project.security.TokenProvider;
