+++ src/main/java/com/sparta/project/service/UserService.java
@@ -5,7 +5,7 @@ import com.sparta.project.dto.match.UserListInMatchDto;
 import com.sparta.project.dto.message.MessageResponseDto;
 import com.sparta.project.dto.user.EvaluationDto;
 import com.sparta.project.dto.user.MyPageResponseDto;
-import com.sparta.project.model.*;
+import com.sparta.project.entity.*;
 import com.sparta.project.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
